fix: refactor imports
- remove dead code
- remove duplicated type definitions
