+++ .pre-commit-config.yaml
@@ -69,7 +69,7 @@ repos:
     hooks:
       - id: knip
         name: knip (node dependency checks)
-        entry: yarn knip --exclude=exports,types,files
+        entry: yarn knip --exclude=exports
         language: system
         pass_filenames: false
         always_run: true
