+++ packages/ansible-language-server/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "declaration": true,
     "outDir": "./out/server",
     "resolveJsonModule": true
   },
