--- src/interfaces/yaml.ts
@@ -1,4 +1,0 @@
-export interface IParsedYaml {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [key: string]: any;
-}
