+++ test/unit/lightspeed/providers/factory.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { LLMProviderFactory } from "../../../../src/features/lightspeed/providers/factory.js";
-import { ProviderType } from "../../../../src/definitions/lightspeed.js";
+import type { ProviderType } from "../../../../src/definitions/lightspeed.d.ts";
 import { PROVIDER_TYPES, TEST_LIGHTSPEED_SETTINGS } from "../testConstants.js";
 
 // Mock AnsibleContextProcessor for providers that extend BaseLLMProvider
